zsh: finally get custom zshrc sourcing
So, long story short, `users.users.${username}.home` needs to be set
(nix-darwin) in order to set
`home-manager.users.${username}.home.homeDirectory`, which in turn is
used to define the `xdg` defaults that are set in `hm-session-vars.sh`.
This script exports some xdg env vars that clobber the custom XDG vars
set in `environment`. Without this home-manager xdg setting, the
`XDG_CONFIG_HOME` var will be set to to `/var/.empty/.config`,
which screws everything up.